# 01-analysis/api/bookings/api.yaml
openapi: 3.1.0
info:
  title: Apex Bookings API
  version: 0.1.0
paths:
  /bookings:
    get:
      operationId: listBookings
      summary: List my bookings
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: Bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: './models.yaml#/components/schemas/Booking'
                  pagination:
                    $ref: '../common/models.yaml#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
    post:
      operationId: createBooking
      summary: Create booking
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models.yaml#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Created booking
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/Booking'
        '409':
          description: Slot full or double booking
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '410':
          description: Slot cancelled
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '422':
          description: Slot already started or validation error
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      summary: Get booking
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '403':
          description: Booking belongs to another client
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
  /bookings/{bookingId}/cancel:
    post:
      operationId: cancelBooking
      summary: Cancel booking
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/Booking'
        '409':
          description: Already cancelled
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '422':
          description: Slot started
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '403':
          description: Booking belongs to another client
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
  /bookings/{bookingId}/marshal-rating:
    post:
      operationId: rateMarshal
      summary: Rate marshal after a ride
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models.yaml#/components/schemas/CreateMarshalRatingRequest'
      responses:
        '200':
          description: Updated booking with marshal_rating
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '403':
          description: Booking belongs to another client
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '409':
          description: Marshal already rated for this booking
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '422':
          description: Booking not eligible for rating or invalid stars
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
    patch:
      operationId: updateMarshalRating
      summary: Update marshal rating after a ride
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models.yaml#/components/schemas/CreateMarshalRatingRequest'
      responses:
        '200':
          description: Updated booking with marshal_rating
          content:
            application/json:
              schema:
                $ref: './models.yaml#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '403':
          description: Booking belongs to another client
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '404':
          description: Booking or rating not found
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
        '422':
          description: Booking not eligible for rating or invalid stars
          content:
            application/json:
              schema:
                $ref: '../common/models.yaml#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      $ref: '../common/models.yaml#/components/securitySchemes/bearerAuth'